2014-04-25 Roman - DB.AGLKS.St.db: KeepAlive timeout of input transport OPC_UA set to 60. - !!!!:FIX:Archive,Archive.*,Special.FLibSYS: Significant values archiving fixes:     Prevent for set negative period values into buffer.     TValBuf::vOK() expand for check aligned to the buffer period.     Disable pass for setVals by lastGet() for direct write allow.     setVals() is adapted for allow write unaligned border [beg:end] to period by round here. - !!!!:DAQ.{BFN,DAQGate,MMS,ModBus,OPC_UA}: Alarms for connection state to remote sources is made unified. - DAQ.MMS: Values read errors processing is added. - UI.QTCfg: rows and collumns of a table adjust mechanism set like from UI.Vision primitive FormEl view "Table".
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2094 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/special/FLibSYS/varchfnc.h
@@ -1,7 +1,7 @@
 
 //OpenSCADA system module Special.FLibSYS file: varchfnc.h
 /***************************************************************************
- *   Copyright (C) 2009 by Roman Savochenko                                *
+ *   Copyright (C) 2009-2014 by Roman Savochenko                           *
  *   rom_as@oscada.org, rom_as@fromru.com                                  *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
@@ -39,10 +39,10 @@ class VArchObj : public TVarObj
 	VArchObj( );
 	~VArchObj( );
 
-	string objName( )       { return "arch"; }
+	string objName( )	{ return "arch"; }
 
 	bool open( const string &inm );
-	bool open( TFld::Type vtp, int isz, int ipr, bool ihgrd, bool ihres );
+	bool open( TFld::Type vtp, int isz, int64_t ipr, bool ihgrd, bool ihres );
 	void close( );
 
 	bool isArch( )		{ return mIsArch && mArch; }
